+++ package.json
@@ -10,6 +10,7 @@
     "packages/remix-cloudflare-pages",
     "packages/remix-cloudflare-workers",
     "packages/remix-deno",
+    "packages/remix-deno-deploy",
     "packages/remix-dev",
     "packages/remix-eslint-config",
     "packages/remix-express",
@@ -25,10 +26,10 @@
     "build": "rollup -c && tsc -b",
     "clean": "git clean -fdX .",
     "format": "prettier --ignore-path .eslintignore --write ./ && npm run lint:fix",
-    "format:deno": "deno fmt integration/helpers/deno-template packages/remix-deno templates/deno --ignore=packages/remix-deno/node_modules",
+    "format:deno": "deno fmt integration/helpers/deno-template packages/remix-deno packages/remix-deno-deploy templates/deno --ignore=packages/remix-deno/node_modules",
     "license": "node ./scripts/license.js",
     "lint": "eslint --cache --ext .tsx,.ts,.js,.jsx,.md .",
-    "lint:deno": "deno lint integration/helpers/deno-template packages/remix-deno templates/deno --ignore=packages/remix-deno/node_modules",
+    "lint:deno": "deno lint integration/helpers/deno-template packages/remix-deno packages/remix-deno-deploy templates/deno --ignore=packages/remix-deno/node_modules",
     "lint:fix": "yarn lint --fix",
     "playground:new": "node ./scripts/playground/new.js",
     "publish": "node ./scripts/publish.js",
