+++ packages/remix-deno-deploy/.empty.js
@@ -1,0 +1,6 @@
+/*
+ Intentionally left empty as a dummy input for Rollup.
+
+ This package should not be bundled by Rollup as its source
+ code is a Deno module, not an NPM package.
+*/
