+++ packages/remix-deno-deploy/README.md
@@ -1,0 +1,14 @@
+# Welcome to Remix!
+
+[Remix](https://remix.run) is a web framework that helps you build better
+websites with React.
+
+To get started, open a new shell and run:
+
+```sh
+npx create-remix@latest
+```
+
+Then follow the prompts you see in your terminal.
+
+For more information about Remix, [visit remix.run](https://remix.run)!
