+++ packages/remix-deno-deploy/globals.ts
@@ -1,3 +1,6 @@
+// TODO: Should come from `@remix-run/deno`
+//  If we would do that now, we would create a circular dependency
+
 /*
 Remix provides `process.env.NODE_ENV` at compile time.
 Declare types for `process` here so that they are available in Deno.
@@ -13,3 +16,5 @@ interface Process {
 var process: Process;
 
 export const installGlobals = () => undefined;
+
+installGlobals();
