+++ packages/remix-deno-deploy/index.ts
@@ -1,0 +1,6 @@
+export type { GetLoadContextFunction, RequestHandler } from "./server";
+export {
+  createRequestHandler,
+  createRequestHandlerWithStaticFiles,
+  serveStaticFiles,
+} from "./server";
