+++ packages/remix-deno/package.json
@@ -14,8 +14,8 @@
   "license": "MIT",
   "sideEffects": false,
   "dependencies": {
-    "@remix-run/server-runtime": "1.6.0",
-    "mime": "^3.0.0"
+    "@remix-run/deno-deploy": "1.6.0",
+    "@remix-run/server-runtime": "1.6.0"
   },
   "engines": {
     "node": ">=14"
