+++ templates/deno/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@remix-run/deno": "*",
+    "@remix-run/deno-deploy": "*",
     "@remix-run/react": "*",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
