+++ templates/deno/server.ts
@@ -1,5 +1,6 @@
 import { serve } from "https://deno.land/std@0.128.0/http/server.ts";
-import { createRequestHandlerWithStaticFiles } from "@remix-run/deno";
+
+import { createRequestHandlerWithStaticFiles } from "@remix-run/deno-deploy";
 // Import path interpreted by the Remix compiler
 import * as build from "@remix-run/dev/server-build";
